requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv